Failed attempt to add an update mode
The problem that add_mapping() can currenly only append to the last
found mapping, instead of to the current end of the list.

Fixing this will require re-writing list insertion. This has to be done
later.

Also add a patch to relocate /etc/inittab to an arbitrary (yet static)
path location.