Short-lived attempt to specify relative order
The idea was to keep a file "order" around instead of file "series"
which includes partial ordering specifications for patches, allowing
"series" to be created on-the-fly.

However, this turned out to be quite useless, because in reality the
order of patches depends on the files touched by the patch.

Therefore, "series" combined with "lsdiff" is good enough to figure out
the dependencies if a new patch needs to be inserted at the right
position within the patch queue.

Nevertheless, checking in "order" because it specified the idea of a
persistent "patch ID" which might become useful in the future.

Also drop patch for "ginit" BusyBox applet from the queue. This applet
does nothing which could not be implemented by a few normal lines of
script code, and only glorifies systemd which certainly needs no
glorification but rather damnation and possible banning.