Add applet symlinks to staging directory
The script dynamically discovers which applets are already available as
"real" commands, skipping symlink creation in this case.

Otherwise a symlink for the applet is created as a command.

This is helpful for uninstallation, because the symlinks can be removed
together with the other installed files, rather than requiring to scan
for and remove dead symlinks.