Better support for ./busybox being a symlink
So far, the scripts assumed to be called with a "../"-prefix. While this
works fine for real directory, it does not work at all for symlinks.

Change the scripts so that they require an additional argument, being
the path to the top-level directory of the BusyBox source files.

The scripts can then be called from the directory containing the
scripts.