Migrate patches to upstream version 1.31.0
Compensate for BusyBox syslogd's new -t option.

This new upstream option ignores timestamps reported by the log sources,
and uses locally-generated new timestamps instead.

This is exactly what the primary function of the syslogd-recvtime patch
has been.

Therefore remove that part of the patch and rename the patch
accordingly.

Another change introduced by upstream is that the applet help texts are
no longer indented with additional spaces. Change the indentation of the
patched help texts in order to comply with this new scheme.