Add script for deriving a static build .config
The new script config_static_patcher.sh reads a normalized Busybox
configuration as created by ./config_normalizer from standard input and
writes a modified version to standard output which will build a
statically-linked shell instead of a dynamically-linked one.

This is intended to be used for rescue-shell versions of Busybox.