Support staging of a static busybox executable
If busybox.static exists in the build directory, it will be included
(properly renamed) into the staging directory, too.

This is intended to be the result of a 2-pass build operation, where
both a static and a dynamically-linked version are built, and the static
version is manually renamed into the beforementioned name.

After this, the staging script can then be run. Without the presence of
the beformentioned file, just the dynamically-linked version will be
staged.