Update 'vi' applet documentation
Add more applets, such as mkfs & fsck for less commonly used
filesystems.

Add a "-g" option the the staging populator which creates a generic (but
of course still plattform-dependent) stage.

Also enhance build support for staging a standalone executable.

Separate this from the filter for just building a statically-linked
version of BusyBox, which is not the same thing.