Add script for building and staging
The script does everything: applies patches, builds 3 different versions
of BusyBox based on a given normalized configuration, and creates a
staging directory containing the result, ready for installation (after
"chown -R root:root"-ing it).

The staging script has also been extended in order to create a
statically-linked rescue shell to be invoked as "/bin/sh" when "/boot"
is mounted as "/". This will also work if "/boot" is a VFAT filesystem,
because it is supposed to also work on "Raspberry Pi" single-board
computers which boot from a FAT-16 formatted SD/SDXC flash memory card.