Install a static version to /bin by default
Option -R has been added to revert to the previous behaviour of
installing the statically-linked version into /usr/local/bin.

Also the build_and_stage_all.sh wrapper script has been expanded to pass
through any -s, -g and -R options.

Regarding -g, this is still auto-detected if the name of the passed
configuration starts with "-all". However, it now is also possible to
explicitly enforce it by passing that option.

The -s will create symlinks to all busybox applets for which there is no
matching command in $PATH at the time when the installation is done.

This was the old standard behaviour which has been made optional now
because it is considered somewhat dangerous.